SA: OpenCV 4 error naming convention
In newer versions of OpenCV (4.x and above), the error codes are defined in the cv::Error enumeration with a slightly different naming convention. Instead of using the CV_ prefix, the error codes are now defined with the Error:: prefix. To fix the error, you can replace CV_StsBadArg with cv::Error::StsBadArg in your code.